Configuring the step to use TLSv1.2 to connect to the TestFairy APIs.
SSLv3 is a really old version of the SSL protocol. It has also proven to be
vulnerable to attacks. As such, most sensible API providers have disabled it on
their servers. This includes TestFairy.

As such, this PR changes the SSL/TLS version used to TLSv1.2. This is the
current version of the TLS protocol.